+++ spec/unit/compose_spec.sh
@@ -1,6 +1,7 @@
 #shellcheck shell=sh
 Include lib/str.sh
 Include lib/core.sh
+Include lib/var.sh
 Include lib/tags.sh
 Include lib/config.sh
 Include lib/compose.sh