+++ spec/unit/var_default_helper.sh
@@ -1,0 +1,3 @@
+#shellcheck shell=sh
+myos_default BUILT 'printf "built-%s" here'
+myos_var BUILT