+++ spec/unit/var_loop_helper.sh
@@ -1,0 +1,3 @@
+#shellcheck shell=sh
+myos_default_LOOPY() { myos_var LOOPY; }
+myos_var LOOPY